+++ src/ConvertFromRawStringCommand.cs
@@ -8,15 +8,17 @@ namespace Sashimi;
 [Alias("a2b")]
 public sealed class ConvertFromRawStringComand : RawCommandBase
 {
-    [Parameter(Mandatory = true, ValueFromPipeline = true, Position = 0)]
+    [Parameter(Mandatory = true, ValueFromPipeline = true, Position = 0,
+               HelpMessageBaseName = MessageBaseName, HelpMessageResourceId = "ConvertFromRawString.parameters.InputString")]
     [AllowEmptyString]
     public string InputString { get; set; } = null!;
 
-    [Parameter()]
+    [Parameter(HelpMessageBaseName = MessageBaseName, HelpMessageResourceId = "ConvertFromRawString.parameters.Encoding")]
+    [ArgumentCompleter(typeof(EncodingCompleter))]
     [Alias("e")]
     public string Encoding { get; set; } = "utf-8";
 
-    [Parameter()]
+    [Parameter(HelpMessageBaseName = MessageBaseName, HelpMessageResourceId = "ConvertFromRawString.parameters.Delimiter")]
     [AllowEmptyString]
     [AllowNull]
     [Alias("d")]
@@ -30,8 +32,15 @@ public sealed class ConvertFromRawStringComand : RawCommandBase
 
     protected override void BeginProcessing()
     {
-        _encoding = System.Text.Encoding.GetEncoding(Encoding);
-        WriteVerboseRaw($"Set encoding: {_encoding.BodyName} [{_encoding.EncodingName}]");
+        try
+        {
+            _encoding = EncodingCompleter.GetEncoding(Encoding);
+        }
+        catch(Exception ex)
+        {
+            ThrowTerminatingError(new(ex, "InvalidEncoding", ErrorCategory.InvalidArgument, this));
+        }
+        WriteVerboseRaw($"Set encoding: {_encoding.WebName} [{_encoding.EncodingName}]");
     }
 
     protected override void ProcessRecord()